+++ src/main/java/net/silentchaos512/lib/command/internal/TeleportCommand.java
@@ -12,7 +12,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
-import net.silentchaos512.lib.util.DimensionId;
 import net.silentchaos512.lib.util.TeleportUtils;
 
 public final class TeleportCommand {
@@ -33,12 +32,12 @@ public final class TeleportCommand {
 
     private static int run(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
         BlockPos target = BlockPosArgument.getLoadedBlockPos(context, "pos");
-        ServerLevel world = DimensionArgument.getDimension(context, "dimension");
+        ServerLevel level = DimensionArgument.getDimension(context, "dimension");
 
         for (Entity entity : EntityArgument.getEntities(context, "entity")) {
             if (entity instanceof Player)
-                TeleportUtils.teleport((Player) entity, DimensionId.fromWorld(world), target.getX(), target.getY(), target.getZ(), null);
-            TeleportUtils.teleportEntity(entity, world, target.getX(), target.getY(), target.getZ(), null);
+                TeleportUtils.teleport((Player) entity, level.dimension(), target.getX(), target.getY(), target.getZ(), null);
+            TeleportUtils.teleportEntity(entity, level, target.getX(), target.getY(), target.getZ(), null);
         }
 
         return 1;